requirements.txt
re
urllib
os
xml.etree.ElementTree
requests
tarfile
collections
matplotlib
shutil
bs4
numpy
json
tqdm
nltk
llm2vec==0.1.8
bitsandbytes
faiss-gpu
packaging
trl
tqdm
captum
seaborn
rouge_score
Levenshtein
matplotlib
wordcloud
bert-score
tensorboard==2.12.0
sentence-transformers==3.0.1
flash_attn
textstat
pyphen
accelerate
time
datasets
evaluate